user joining or leaving now shows in chat, chat now auto-scroll if a new message is received
